Terminando a o style do calendário e o modalCriarProjeto
