+++ src/App.css
@@ -1,3 +1,12 @@
+.sx-react-calendar-wrapper {
+  width: 100%;
+  height: 100vh;
+  max-height: 90vh;
+}
+  .sx__chevron-wrapper{
+    border-radius: 10px;
+  }
+
 .App {
   text-align: center;
 }
