+++ src/Componets/Modal/ModalCadastroProjeto.jsx
@@ -2,6 +2,7 @@ import { useState } from "react";
 import Modal from "react-modal";
 import ApiService from "../../services/ApiService";
 import ToastService from "../../services/ToastService";
+import styles from "./ModalCadastroProjeto.module.css";
 
 export default function ModalCadastroProjeto({ modalAberto, setModalAberto }) {
   Modal.setAppElement("#root");
@@ -20,12 +21,14 @@ export default function ModalCadastroProjeto({ modalAberto, setModalAberto }) {
   const [descricao, setDescricao] = useState("");
 
   async function CadastrarProjeto() {
+    console.log("test");
     try {
       const body = {
         nome,
         descricao,
       };
 
+      console.log(body);
       await ApiService.post("/Projeto/CriarProjeto", body);
       ToastService.Success("Projeto Criado com Sucesso");
       setModalAberto(false);
@@ -36,27 +39,34 @@ export default function ModalCadastroProjeto({ modalAberto, setModalAberto }) {
 
   return (
     <Modal isOpen={modalAberto} style={customStyles}>
-      <h2>Criar Projeto</h2>
-      <button
-        onClick={() => {
-          setModalAberto(false);
-        }}
-      >
-        Fechar
-      </button>
+      <div className={styles.container}>
+        <div className={styles.sidebar}>
+          <h3 className={styles.title}>Vamos Criar um Projeto</h3>
 
-      <input
-        placeholder="Nome"
-        value={nome}
-        onChange={(e) => setNome(e.target.value)}
-      />
-      <input
-        placeholder="Descrição"
-        value={descricao}
-        onChange={(e) => setDescricao(e.target.value)}
-      />
-
-      <button>Criar Projeto</button>
+          <p>Nome do Projeto</p>
+          <input
+            placeholder="Nome"
+            value={nome}
+            onChange={(e) => setNome(e.target.value)}
+          />
+          <br />
+          <p>Descrição do Projeto</p>
+          <input
+            placeholder="Descrição"
+            value={descricao}
+            onChange={(e) => setDescricao(e.target.value)}
+          />
+          <br />
+          <center>
+            <button className={styles.button} onClick={CadastrarProjeto}>
+              Criar Projeto
+            </button>
+          </center>
+        </div>
+        <div className={styles.right}>
+          <p>Crie um projeto para gerenciar e organizar suas tarefas </p>
+        </div>
+      </div>
     </Modal>
   );
 }
