+++ src/Componets/Modal/ModalCadastroProjeto.module.css
@@ -1,0 +1,41 @@
+.container {
+  display: flex;
+  flex-direction: row;
+  height: 100%;
+  width: 100%;
+  background-color: #222831;
+}
+
+.sidebar {
+  margin: 30px;
+  justify-content: center;
+  align-items: center;
+  height: 100%;
+  width: 30%;
+  background-color: #222831;
+  color: white;
+}
+.right {
+  color: white;
+  width: 70%;
+  background-color: blue;
+}
+
+input {
+  color: white;
+  width: 200px;
+  margin-bottom: 10px;
+  padding: 10px;
+  border-radius: 10px;
+  background-color: #7b7b7b;
+}
+p {
+  margin-bottom: 10px;
+}
+.button {
+  background-color: #7b7b7b;
+}
+.title {
+  width: 300px;
+  margin-bottom: 10px;
+}
