+++ src/pages/Calendario/Calendario.module.css
@@ -1,0 +1,15 @@
+.sx-react-calendar-wrapper {
+  width: 100%;
+  height: 800px;
+  max-height: 90vh;
+}
+
+.sx__chevron-wrapper {
+    position: relative;
+    border-radius: 10px;
+    min-height: 48px;
+    min-width: 48px;
+    cursor: pointer;
+    transition: background-color 0.2s ease-in-out;
+    font-size: 0;
+}
