+++ src/pages/Home/Home.jsx
@@ -1,7 +1,11 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import ListaDeProjetos from "../../Componets/ListaDeProjetos/ListaDeProjetos";
+import AuthService from "../../services/AuthServices";
+import ApiService from "../../services/ApiService";
+import { useNavigate } from "react-router-dom";
 
 export default function Home() {
+  const navigate = useNavigate();
   const [projetos, setProjetos] = useState([
     {
       tituloTarefa: "Banco de Dados",
@@ -10,6 +14,17 @@ export default function Home() {
       descricaoProjeto: "TaskSync",
     },
   ]);
+  useEffect(() => {
+    VerificarLogin();
+  }, []);
+
+  function VerificarLogin() {
+    const usuarioEstaLogado = AuthService.VerificarSeUsuarioEstaLogado();
+    if (!usuarioEstaLogado) {
+      navigate("/");
+    }
+  }
+
   return (
     <div>
       <ListaDeProjetos projetos={projetos} />
