--- src/pages/HomeOld/ListaDeProjetos.jsx
@@ -1,6 +1,0 @@
-import Home from "./Home";
-import React, { useState } from "react";
-
-export default function ListaDeProjetos() {
-  return <div></div>;
-}
