--- src/pages/HomeOld/ListadeTarefas.jsx
@@ -1,15 +1,0 @@
-import Home from "./Home";
-import React, { useState } from "react";
-
-export default function ListaDeTarefas() {
-  const [projeto, setProjeto] = useState({
-    tituloTarefa: "Banco de Dados",
-    descricaoTarefa: "AAAAAAAAAAAAAA",
-    tituloProjeto: "TaskSync",
-  });
-  return (
-    <div>
-      
-    </div>
-  );
-}
