+++ src/services/ApiService.js
@@ -11,7 +11,7 @@ function createHeader() {
     };
   }
 }
-const baseUrl = "https://localhost:44378/api";
+const baseUrl = "https://localhost:7121/api";
 const ApiService = {
   async get(endpoint) {
     const header = createHeader();
